scripts/requirements.txt
pyfunda
simple_http_server
